requirements.txt
atproto==0.0.9
requests==2.32.3
Pillow==10.4.0